Enabling BBS message parsing.
Emailing NAVI will allow you to request a list of threads
from the communities you are subscribed to.

Currently, we can retrieve a set of threads.

The next step will be retrieving all the messages from the
thread, and then retrieving all the user profiles in the
thread.  Finally, we will convert that into an email and
send the email to the user.